Network API refactor: rename modules
