+++ bin/cylc-remove
@@ -34,7 +34,7 @@ if '--use-ssh' in sys.argv[1:]:
 import cylc.flags
 from cylc.prompt import prompt
 from cylc.option_parsers import CylcOptionParser as COP
-from cylc.network.client import SuiteRuntimeServiceClient
+from cylc.network.httpclient import SuiteRuntimeServiceClient
 
 
 def main():
