+++ tests/api-suite-info/00-get-graph-raw-1/bin/ctb-get-graph-raw
@@ -21,7 +21,7 @@ import json
 import os
 import sys
 
-from cylc.network.client import SuiteRuntimeServiceClient
+from cylc.network.httpclient import SuiteRuntimeServiceClient
 
 
 def main():
